requirements.txt
asttokens==2.0.5
colorama==0.4.5
cycler==0.10.0
decorator==5.1.1
executing==0.9.0
google-play-scraper==1.2.2
ipython-genutils==0.2.0
Jinja2==3.1.2
jupyterlab-widgets==1.1.1
kiwisolver==1.3.1
MarkupSafe==2.1.1
matplotlib==3.3.4
mysql-connector-python==8.0.28
numpy==1.20.1
opencv-python==4.5.1.48
pandas==1.4.3
pickleshare==0.7.5
Pillow==8.3.2
prompt-toolkit==3.0.30
pure-eval==0.2.2
Pygments==2.12.0
pyparsing==2.4.7
PySastrawi==1.2.0
pytesseract==0.3.7
python-dateutil==2.8.1
pytz==2022.1
pywin32==304
six==1.15.0
stack-data==0.3.0
tornado==6.2
traitlets==5.3.0
wcwidth==0.2.5
wordcloud==1.8.2.2
